requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
sqlalchemy==2.0.28
pandas==2.2.1
openpyxl==3.1.2
langchain==0.3.27
langchain-community==0.3.31
langchain-google-genai==2.1.12
python-dotenv==1.0.1
sqlparse==0.4.4